# application/api-spec/src/main/resources/openapi/isk.yaml
openapi: 3.0.0
info:
  title: Informačný systém knižnice (ISK)
  version: 1.0.0
  description: API pre správu knižničného systému – katalóg kníh, výpožičky, rezervácie a členstvá.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /members:
    get:
      summary: Získanie zoznamu všetkých členov
      operationId: getAllMembers
      tags:
        - Member
      responses:
        '200':
          description: Zoznam členov
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Vytvorenie nového člena
      operationId: createMember
      tags:
        - Member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberRequest'
      responses:
        '201':
          description: Člen úspešne vytvorený
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /members/{id}:
    get:
      summary: Získanie člena podľa ID
      operationId: getMemberById
      tags:
        - Member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detail člena
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /members/{id}/membership/renew:
    post:
      summary: Obnovenie členstva
      operationId: renewMembership
      tags:
        - Member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Členstvo úspešne obnovené
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /members/{id}/fines/{fineId}/pay:
    post:
      summary: Zaplatenie pokuty
      operationId: payFine
      tags:
        - Member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: fineId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pokuta úspešne zaplatená
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /members/{id}/fines/{fineId}/waive:
    post:
      summary: Odpustenie pokuty
      operationId: waiveFine
      tags:
        - Member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: fineId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pokuta úspešne odpustená
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── LOANS ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    Money:
      type: object
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: EUR

    # ─── BOOK ───────────────────────────────────────────────────────────────
...
    Member:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        memberRole:
          $ref: '#/components/schemas/MemberRole'
        membership:
          $ref: '#/components/schemas/Membership'
        fines:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Fine'
    CreateMemberRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - memberRole
      properties:
        firstName:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        lastName:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        email:
          type: string
          format: email
        memberRole:
          $ref: '#/components/schemas/MemberRole'
    Membership:
      type: object
      properties:
        expiryDate:
          type: string
          format: date
          readOnly: true
        status:
          $ref: '#/components/schemas/MembershipStatus'
    Fine:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        amount:
          $ref: '#/components/schemas/Money'
        reason:
          type: string
        status:
          $ref: '#/components/schemas/FineStatus'
    MemberRole:
      type: string
      enum:
        - MEMBER
        - LIBRARIAN
        - ADMIN
    MembershipStatus:
      type: string
      enum:
        - ACTIVE
        - EXPIRED
        - SUSPENDED
    FineStatus:
      type: string
      enum:
        - PENDING
        - PAID
        - WAIVED

    # ─── LOAN ───────────────────────────────────────────────────────────────
...